Move streaming into a separate avml-stream binary
Snapshot streaming is a different workflow from acquire-then-upload —
it picks one memory source up front (no auto-retry), never touches
local disk, and has its own destination protocol. Putting it in a
separate binary mirrors how avml-convert and avml-upload are handled,
and keeps the avml CLI flat (no acquire/stream subcommand split).

avml: revert to the pre-streaming flat CLI shape. The --stream-to-blob
flag and its handler move out; --max-disk-usage, --sas-url, --delete
etc. remain unchanged. The mandatory positional <FILENAME> argument
is now always a plain PathBuf, no more cfg-gated Option<PathBuf>.

avml-stream: new binary, requires the blobstore feature. Subcommand
shape:

  avml-stream blob <SAS_URL> [--compress] [--source S]
                             [--sas-block-size N]
                             [--sas-block-concurrency N]

The subcommand layer leaves room for adding additional destination
protocols later (e.g., a future 'avml-stream s3-multipart' against
the AWS SDK) without restructuring.

Cargo.toml: register the new [[bin]] with required-features =
['blobstore']. eng/build.sh and the linux artifact upload paths in
build.yml include the new binary. README points at avml-stream for
the no-local-file streaming workflow.

Verified: fmt, full pedantic clippy, all tests (36 lib + 2 convert +
1 doctest), semver-checks (additive only, no version bump), musl
release builds with --all-features and --no-default-features.
--help output of avml is unchanged from main; avml-stream's --help
shows the new subcommand surface.
